Fix Lutris GUI reporting Vita installed titles as missing
The Vita3K emulator launches Vita titles using their Title ID and not a file path.
Therefore the Vita3K runner is setup to accept the title ID field.
However that title ID field was set as the 'main_file' option in the config yaml that gets generated when a Vita entry is added.
The missing installed title check in Lutris invokes the game.py [get_path_from_config](https://github.com/lutris/lutris/blob/4620e1ba2fafb90ac1eef99dad4e5f65017d8554/lutris/game.py#L643) which treats the `main_file` key as a file path.

As the Vita title ID is not a valid file path, the Missing Games check would flag all Vita installed titles as missing.

The fix here is to change the key from 'main_file' to 'title_id' so that the field is not treated as a file path.