Added Open Folder Location for game art
Refactored the game_common class to split out the Game Info panel UI to it's own file.
The Game Info panel has been updated to support the "Advanced Options" switch.
When the Advanced Options switch is selected, several text entries become available with the path to the art images on the filesystem, as well as a Open Folder location button to jump to the location of the art image.

This is the first pass at implementing the UI mocks from issue 4213 to select cover art from local or online sources