Fix runner.get_executable method when filename changes
The `runner.install` method downloads the runner executable into the directory of \<RUNNER_DIR>/\<runner-filename>
However the `runner.get_executable` method checks the runner executable within a directory of \<RUNNER_DIR>/\<runner-executable-field>.

What this means is if a JSON runner file has a name of "foo-emu.json" and the "runner_executable" property inside of it is set to "**foo**/foo.exe", then the `get_executable` method will never locate the downloaded runner.

To workaround, this issue, the `runner_get_executable` method now also checks for the path of \<RUNNER_DIR>/\<runner-filename>/\<runner-executable> to see if the executable resides within that directory.
With the workaround, the subfolder in front of the actual executable path can be removed.
i.e the "runner_executable" property for a should now be set to "foo.exe", instead of "**foo**/foo.exe"

With this change the `get_executable` method will not be out of sync with the `download` method.