Converted Vita3K runner to a hybrid JSON/Python runner
The Vita3K runner field definition now resides in the vita3k.json file.
The Python file contains a custom play() method which maps runner old entry_point_option field of "main_file" to the new "title_id" entry point since the Vita3K doesn't use file paths to run.

This commit is to be an example of how to define a runner in JSON and then use Python to extend that runners functionality.